Change Single/Multiple graph display buttons
